requirements.txt
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
backrefs==5.8
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
ghp-import==2.1.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Markdown==3.8
MarkupSafe==3.0.2
mergedeep==1.3.4
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pymdown-extensions==10.15
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyyaml_env_tag==0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
